oauth2 service factored using generics
