+++ PhotoGrammy/Services/URLRequest+Extensions.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+extension URLRequest {
+    static func makeHTTPRequest(path: String,
+                                httpMethod: String,
+                                baseURL: URL = DefaultBaseURL) -> URLRequest {
+        var request = URLRequest(url: URL(string: path, relativeTo: baseURL)!)
+        request.httpMethod = httpMethod
+        return request
+    }
+}
