+++ PhotoGrammy/Services/URLSession+Extensions.swift
@@ -1,0 +1,43 @@
+import Foundation
+
+extension URLSession {
+    
+    enum NetworkError: Error {
+        case httpStatusCode(Int)
+        case urlRequestError(Error)
+        case urlSessionError
+    }
+    
+    func objectTask<T: Decodable>(
+        for request: URLRequest,
+        completion: @escaping (Result<T, Error>) -> Void
+    ) -> URLSessionTask {
+        
+        let fulfillCompletion: (Result<T, Error>) -> Void = { result in
+            DispatchQueue.main.async{
+                completion(result)
+            }
+        }
+        
+        let task = dataTask(with: request) { data, response, error in
+            
+            if let data = data,
+               let response = response,
+               let statusCode = (response as? HTTPURLResponse)?.statusCode {
+                if 200 ..< 300 ~= statusCode {
+//                    guard let self = self else { return }
+                    JSONDecoder().decode(data: data, completion: fulfillCompletion)
+                } else {
+                    fulfillCompletion(.failure(NetworkError.httpStatusCode(statusCode)))
+                }
+            } else if let error = error {
+                fulfillCompletion(.failure(NetworkError.urlRequestError(error)))
+            } else {
+                fulfillCompletion(.failure(NetworkError.urlSessionError))
+            }
+        }
+        
+        task.resume()
+        return task
+    }
+}
